Bug: #1435 - Visible rectangle calculation in Image Renderer is incorrect Bug: #502 - General defect for missing Unit Tests
- Restored/fixed bilinear interpolation unit tests.
- Some minor changes had to be made to the rendering code because the source/destination rectangles were being calculated incorrectly.  The destination rectangle was a RectangleF which would occasionally cause an off-by-one error in the resulting bitmap, such that it would be shifted in one direction by a pixel.  For example, if the destination rectangle were calculated to be at (1.9999, 2), then the interpolator would start filling in the bitmap at pixel (1, 2) instead of (2, 2) essentially causing the interpolated image to be shifted by a pixel.
- Although the resulting image won't look any different to the user, the calculation was incorrect - the destination rectangle, since it is a screen bitmap, should be specified as an integer rectangle so that there is no confusion as to which pixels are to be filled in by the interpolator.
- If this change hadn't been made, it would not have been possible to write correct unit tests without slightly fudging the numbers, and this way the unit tests can remain automated.
- The bilinear interpolator and the unit tests now agree to within one-one-thousandth of a source pixel for various flips/scales/rotations/translations.
- Image display for all image types and bit depths should be tested thoroughly.

+++ ImageViewer/Rendering/BilinearInterpolation/BilinearInterpolation.h
@@ -42,10 +42,10 @@ extern "C"
             unsigned int dstWidth,
             unsigned int dstBytesPerPixel,
 
-			float dstRegionRectLeft,
-            float dstRegionRectTop,
-            float dstRegionRectRight,
-            float dstRegionRectBottom,
+			int dstRegionRectLeft,
+            int dstRegionRectTop,
+            int dstRegionRectRight,
+            int dstRegionRectBottom,
 
 			BOOL swapXY,
 			LUTDATA* pLutData
